// pork-ionic-angular/src/app/interfaces/ronda-interface.ts
export class RondaInterface {
    year= new Date().getFullYear();
    fecha : string = '';
    ultimoDia: string ='';
    semana: string = '';
    producto: string ='';
    cantidad: number= 0;
    peso: number =0;
    precio : number = 0;     
    mercado: string = '';    
    comentario: string ='';
    entrega: string= '';
    usuario: string='';
}